pyproject.toml
[project]
name = "manual-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "google-genai>=1.56.0",
    "google-generativeai>=0.8.6",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
